changed drone id to 64 bits to fix struct alignment. Added comments to main.h

+++ src/drone.c
@@ -11,7 +11,7 @@
 #define MAX_ANGLE 0.01
 #include "flightplan.h"
 
-static uint32_t ids = 0;
+static uint64_t ids = 0;
 
 inline double generateGaussian(double mean, double stdDev)
 {
@@ -67,7 +67,6 @@ Drone DR_newDrone(double x, double y, double vx, double vy, double size)
 	d.velocity.x = vx;
 	d.velocity.y = vy;
 	d.fp = FP_newFlightPlan(4);
-	// FP_push_waypoint(d.fp,d.position);
 	d.size = size;
 	return d;
 }
